refactor: Splitted db-client into collection-interfaces
